requirements.txt
fastapi
opencv-python
easyocr
numpy
selenium
uvicorn
requests
python-telegram-bot